cleanup dedup code
